+++ Restaurant/Ingredient.cs
@@ -7,7 +7,7 @@ namespace Restaurant
 		public string Name { get; set; }
 		public double Price { get; set; }
 
-		public Ingredient (string name, float price)
+		public Ingredient (string name, double price)
 		{
 			Name = name;
 			Price = price;
